% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_portfolio.R
\name{export.portfolio}
\alias{export.portfolio}
\title{export.portfolio}
\usage{
export.portfolio(portfolio, path)
}
\arguments{
\item{portfolio}{portfolio as created by Portfolio.Manager}

\item{path}{Location on the disk where the portfolio should be saved}
}
\value{
Export portfolio into a csv file
}
\description{
Exports the portfolio into a csv file
}
\details{
Export your portfolio
}
