% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_1_month_category_ret.R
\name{ms.1mCategoryReturn}
\alias{ms.1mCategoryReturn}
\title{ms.1mCategoryReturn}
\usage{
ms.1mCategoryReturn(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER}
}
\value{
The output gives the category return of the funs
}
\description{
Provides the 1 month Category Return of the Security
}
\details{
Gives the category return for the input TICKER (Fund)
}
\examples{
ms.1mCategoryReturn('FXAIX')
}
