% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_3y_alpha.R
\name{ms.3yAlpha}
\alias{ms.3yAlpha}
\title{ms.3yAlpha}
\usage{
ms.3yAlpha(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER}
}
\value{
The output gives the 3-Year Alpha of the funs
}
\description{
Provides the Total Return of the Security
}
\details{
Gives the 3-Year Alpha for the input TICKER (Fund)
}
\examples{
ms.3yAlpha('FXAIX')
}
