% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_3y_category_ret.R
\name{ms.3yCategoryReturn}
\alias{ms.3yCategoryReturn}
\title{ms.3yCategoryReturn}
\usage{
ms.3yCategoryReturn(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER}
}
\value{
The output gives the category return of the funs
}
\description{
Provides the Category Return of the Security
}
\details{
Gives the category return for the input TICKER (Fund)
}
\examples{
ms.3yCategoryReturn('FXAIX')
}
