% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_3y_excess_cat_return.R
\name{ms.3yExcessCategoryReturn}
\alias{ms.3yExcessCategoryReturn}
\title{ms.3yExcessCategoryReturn}
\usage{
ms.3yExcessCategoryReturn(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER}
}
\value{
The output gives the Excess category return of the funs
}
\description{
Provides the Excess Category Return of the Security
}
\details{
Gives the Excess category return for the input TICKER (Fund)
}
\examples{
ms.3yExcessCategoryReturn('FXAIX')
}
