% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_3y_rank.R, R/ms_5y_rank.R
\name{ms.3yRank}
\alias{ms.3yRank}
\alias{ms.5yRank}
\title{ms.3yRank}
\usage{
ms.3yRank(ticker)

ms.5yRank(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER}
}
\value{
The output gives the stated rank in caegory of the funs

The output gives the stated rank in caegory of the funs
}
\description{
Provides the 3y Rank of the Security

Provides the 5y Rank of the Security
}
\details{
Gives the stated rank for the input TICKER (Fund)

Gives the stated rank for the input TICKER (Fund)
}
\examples{
ms.3yRank('FXAIX')
ms.3yRank('FXAIX')
}
