% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_3y_return.R
\name{ms.3yReturn}
\alias{ms.3yReturn}
\title{ms.3yReturn}
\usage{
ms.3yReturn(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER}
}
\value{
The output gives the stated return of the funs
}
\description{
Provides the 3y Return of the Security
}
\details{
Gives the stated return for the input TICKER (Fund)
}
\examples{
ms.3yReturn('FXAIX')
}
