% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_5y_excess_cat_return.R
\name{ms.5yExcessCategoryReturn}
\alias{ms.5yExcessCategoryReturn}
\title{ms.5yExcessCategoryReturn}
\usage{
ms.5yExcessCategoryReturn(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER}
}
\value{
The output gives the Excess category return of the funs
}
\description{
Provides the Excess Category Return of the Security
}
\details{
Gives the Excess category return for the input TICKER (Fund)
}
\examples{
ms.5yExcessCategoryReturn('FXAIX')
}
