% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_5y_return.R
\name{ms.5yReturn}
\alias{ms.5yReturn}
\title{ms.5yReturn}
\usage{
ms.5yReturn(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER}
}
\value{
The output gives the stated return of the funs
}
\description{
Provides the Total Return of the Security
}
\details{
Gives the stated return for the input TICKER (Fund)
}
\examples{
ms.5yReturn('FXAIX')
}
