% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fund_name.R
\name{ms.FundName}
\alias{ms.FundName}
\title{ms.FundName}
\usage{
ms.FundName(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER}
}
\value{
The output gives the Fund Name
}
\description{
Provides the Fund Name
}
\details{
Gives the stated TICKER Fund Name
}
\examples{
ms.FundName('FXAIX')
}
