% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_holdings_turnover.R
\name{ms.HoldingsTurnover}
\alias{ms.HoldingsTurnover}
\title{ms.HoldingsTurnover}
\usage{
ms.HoldingsTurnover(ticker)
}
\arguments{
\item{ticker}{Ticker of the Fund}
}
\value{
The output provides the Input Performance Metric as given by MorningStar.
}
\description{
Gives the ouptput of input Performance Metric as provided by the MorningStar Performance Results
}
\details{
Morningstar provides a great deal of useful information in its print publications and in its Principia Plus CD-ROM data service. Use this function for performance metric calculations.
}
\examples{
ms.HoldingsTurnover('FXAIX')
}
