% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_percent_change.R
\name{ms.PercentChange}
\alias{ms.PercentChange}
\title{ms.PercentChange}
\usage{
ms.PercentChange(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER of the fund}
}
\value{
Provides the percent change of the fund
}
\description{
Gives the percent change of the fund
}
\details{
Percent Change of the funds from previous Close.
}
\examples{
ms.PercentChange('FXAIX')
}
