% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_rating.R
\name{ms.Rating}
\alias{ms.Rating}
\title{ms.Rating}
\usage{
ms.Rating(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER}
}
\value{
The output gives the Morningstar Rating of the funs
}
\description{
Provides the Morningstar Rating of the Security
}
\details{
Gives the stated Morningstar Rating for the input TICKER (Fund)
}
\examples{
ms.Rating('FXAIX')
}
