% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_return_rating.R
\name{ms.ReturnRating}
\alias{ms.ReturnRating}
\title{ms.ReturnRating}
\usage{
ms.ReturnRating(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER}
}
\value{
The output gives the Morningstar Return Rating of the funs
}
\description{
Provides the Morningstar Return Rating of the Security
}
\details{
Gives the stated Morningstar Return Rating for the input TICKER (Fund)
}
\examples{
ms.ReturnRating('FXAIX')
}
