% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_risk_rating.R
\name{ms.RiskRating}
\alias{ms.RiskRating}
\title{ms.RiskRating}
\usage{
ms.RiskRating(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER}
}
\value{
The output gives the Morningstar Risk Rating of the funs
}
\description{
Provides the Morningstar Risk Rating of the Security
}
\details{
Gives the stated Morningstar Risk Rating for the input TICKER (Fund)
}
\examples{
ms.RiskRating('FXAIX')
}
