% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_stdev.R
\name{ms.stdev}
\alias{ms.stdev}
\title{ms.stdev}
\usage{
ms.stdev(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER for the security}
}
\value{
Returns the Standard Deviation of the Security as Calculated by MorningStar.
}
\description{
MorningStar 3-yr Standard Deviation
}
\details{
Morningstar's reported standard deviation of fund returns uses both the monthly standard deviation and monthly mean return to compute an annualized value, assuming compounding of the monthly returns and zero serial correlation.
}
\examples{
ms.stdev('AAPL')
}
