% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_summary.R
\name{ms.summary}
\alias{ms.summary}
\title{ms.summary}
\usage{
ms.summary(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER for the security}
}
\value{
Returns the Summary of the Security's Risk and Return as Calculated by MorningStar.
}
\description{
Summary of the Fund Risk and Returns
}
\details{
Morningstar provides a great deal of useful information in its print publications and in its Principia Plus CD-ROM data service. This function can be used to get a DataFrame of the summary of the risk and return of the input fund.
}
\examples{
ms.summary('FXAIX')
}
