% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_top10_holding.R, R/ms_top10_holding_total.R
\name{ms.Top10Holding}
\alias{ms.Top10Holding}
\alias{ms.Top10HoldingTotal}
\title{ms.Top10Holding}
\usage{
ms.Top10Holding(ticker)

ms.Top10HoldingTotal(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER of the Fund}
}
\value{
Gives the Morningstar Top 10 Holding of the Fund

Gives the Morningstar Top 10 Holding Total of the Fund
}
\description{
Gives the MorningStar Top 10 Holding of the fund

Gives the MorningStar Top 10 Holding Total of the fund
}
\details{
Gives the fund's top 10 Holdings as cmputed by MorningStar.

Gives the fund's top 10 Holdings Total as cmputed by MorningStar.
}
\examples{
ms.Top10Holding('FXAIX')
ms.Top10HoldingTotal('FXAIX')
}
