% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_top_10_holding_plot.R
\name{ms.Top10HoldingPlot}
\alias{ms.Top10HoldingPlot}
\title{ms.Top10HoldingPlot}
\usage{
ms.Top10HoldingPlot(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER of the Fund}
}
\value{
Gives the Morningstar Top 10 Holding Plot of the Fund
}
\description{
Gives the MorningStar Top 10 Holding Plot of the fund
}
\details{
Gives the fund's top 10 Holdings Plot as cmputed by MorningStar.
}
\examples{
ms.Top10HoldingPlot('FXAIX')
}
