% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_ytd_category_ret.R
\name{ms.YTDCategoryReturn}
\alias{ms.YTDCategoryReturn}
\title{ms.YTDCategoryReturn}
\usage{
ms.YTDCategoryReturn(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER}
}
\value{
The output gives the category return of the funs
}
\description{
Provides the Category Return of the Security
}
\details{
Gives the category return for the input TICKER (Fund)
}
\examples{
ms.YTDCategoryReturn('FXAIX')
}
