% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_ytd_rank.R
\name{ms.YtdRank}
\alias{ms.YtdRank}
\title{ms.YtdRank}
\usage{
ms.YtdRank(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER}
}
\value{
The output gives the stated rank in caegory of the funs
}
\description{
Provides the YTD Rank of the Security
}
\details{
Gives the stated rank for the input TICKER (Fund)
}
\examples{
ms.YtdRank('FXAIX')
}
