% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_ytdreturn.R
\name{ms.ytdreturn}
\alias{ms.ytdreturn}
\title{ms.ytdreturn}
\usage{
ms.ytdreturn(ticker)
}
\arguments{
\item{ticker}{Enter the TICKER for the security}
}
\value{
Returns the MorningStar YTD Return of the Security as Calculated by MorningStar.
}
\description{
MorningStar YTD Return
}
\details{
A traditional measure of performance in the mutual fund industry is the cumulative value of $1 compounded over a specified number of periods. Morningstar uses such measures on both before-load and after-load fee bases. The value relative at the end of T periods not taking any load charges into account is:
}
\examples{
ms.ytdreturn('AAPL')
}
