% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_manager.R
\name{PortfolioManager}
\alias{PortfolioManager}
\title{PortfolioManager}
\usage{
PortfolioManager(ticker, view = c("overview", "detailed"))
}
\arguments{
\item{ticker}{Enter the Fund TICKER}

\item{view}{Choose a View of the Portfolio, Default: c("overview", "detailed")}
}
\value{
The output is a tibble that gives the overview of the risk and reward metrics of different holdings in a portfoluio.
}
\description{
Gives the detailed  Risk-Reward Metrics as Computed by MorningStar
}
\details{
This function can be used to examine and assess your holdings in different funds
}
